% Generated by roxygen2 (4.0.0): do not edit by hand
\name{checkForGitbook}
\alias{checkForGitbook}
\title{Check if Gitbook is installed.}
\usage{
checkForGitbook(quiet = FALSE)
}
\arguments{
\item{quiet}{logical indicating whether messages should be printed.}
}
\description{
If Gitbook is not installed it will try to do so. If the installion fails or
Node.js is not installed, an error will be thrown.
}

