% Generated by roxygen2 (4.0.0): do not edit by hand
\name{gitbookInfo}
\alias{gitbookInfo}
\title{Returns the version information about the currently installed gitbook and
what is avaialble from \url{https://www.npmjs.org/}.}
\usage{
gitbookInfo()
}
\value{
a character vector with two elements, \code{installed.version} and
        \code{available.version}.
}
\description{
Returns the version information about the currently installed gitbook and
what is avaialble from \url{https://www.npmjs.org/}.
}

