\name{installGitbook}
\alias{installGitbook}
\title{This will install Gitbook.io.}
\usage{
  installGitbook()
}
\description{
  This function will install Gitbook from
  \url{http://nodejs.org/download/} regardless if the
  current version. That is, this function will either
  update or install a new version of Gitbook.io.
}

