% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/newPage.R
\name{newPage}
\alias{newPage}
\title{Generate R markdown file}
\usage{
newPage(filename = "newpage.Rmd", dir = getwd(), title = "Title")
}
\arguments{
\item{filename}{target filename.  Default is 'newpage.Rmd'.}

\item{dir}{target destination of the file.  Default is current working directory.}

\item{title}{title to insert into Rmd file.  Default is 'Title'.}
}
\description{
Generate R markdown file
}
\examples{
newPage()
newPage(filename = "mynewpage")
newPage(filename = "mynewpage.Rmd", title = "A custom title")
}

