\docType{package}
\name{Rgitbook-package}
\alias{Rgitbook-package}
\title{Gitbook Projects with R Markdown}
\description{
Gitbook Projects with R Markdown
}
\details{
This package provides functions for working with
\url{Gitbook.io} projects from within R. Moreover, it
provides support for working with R markdown files using
the \code{knitr} package. Support for MathJax is also
included.
}
\author{
\email{jason@bryer.org}
}
\keyword{Gitbook}
\keyword{Github}
\keyword{Markdown}
\keyword{R}

