% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newGitbook.R
\name{newGitbook}
\alias{newGitbook}
\title{Initializes a new Gitbook.}
\usage{
newGitbook(dir)
}
\arguments{
\item{dir}{location of the built gitbook.}
}
\description{
This will initalize a new Gitbook in the given directory. When done, it will
also change the working directory.
}

